+++ CHANGELOG.md
@@ -31,10 +31,10 @@ All notable changes to this project will be documented in this file. This change
     - `preflex.core`  to `preflex.resilient`
     - `preflex.error` to `preflex.resilient.error`
     - `preflex.impl`  to `preflex.resilient.impl`
-  - [BREAKING CHANGE] Make duration arguments as `preflex.type.IDuration` instead of milliseconds
+  - [BREAKING CHANGE] Overhaul duration and time arguments in `preflex.resilient` namespace
     - In function `preflex.resilient/make-discrete-fault-detector`
-      - [TODO] Change argument `connected-until-duration` to duration instead of milliseconds
-      - [TODO] Drop option `:now-finder` in favor of milliseconds
+      - Change argument `connected-until-duration` to duration instead of milliseconds
+      - Rename option `:now-finder` to `:now-millis-finder` - time calculation in milliseconds
     - In function `preflex.resilient/make-rolling-fault-detector`
       - [TODO] Change argument `connected-until-duration` to duration instead of milliseconds
       - [TODO] Change option `:bucket-interval` from milliseconds to duration